refactor: decompose ChatViewModel into focused extensions
Split the 948-line base class into smaller, focused extension files
by extracting message management, model loading, session management,
and memory pressure concerns into separate files.

New files:
- ChatViewModel+Messages.swift (send, edit, regenerate, stop, clear, export, pin)
- ChatViewModel+ModelLoading.swift (dispatch, load, unload coordination)
- ChatViewModel+SessionManagement.swift (switch session, save settings, model discovery, lifecycle)
- ChatViewModel+MemoryPressure.swift (monitoring and pressure handling)

The base ChatViewModel.swift is reduced from 948 to 382 lines,
containing only stored properties, init, and configure methods.
